% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlFormat.R
\name{apply_insertions}
\alias{apply_insertions}
\title{inserts html tags into a string}
\usage{
apply_insertions(string, insertions, tags.df)
}
\description{
inserts html tags into a string
}
\examples{
string = "Hello World"
insertions <- c("<b>" = 3, "</b>" = 6)
apply_insertions(string, insertions)
}

