% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputAnnotate.R
\name{inputAnnotate}
\alias{inputAnnotate}
\title{Create an input control that returns the positions of the
selected text.}
\usage{
inputAnnotate(inputId, value = "")
}
\description{
Create an input control that returns the positions of the
selected text.
}

