% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlFormat.R
\name{insert_tags}
\alias{insert_tags}
\title{inserts html tags into a string for a data.tree}
\usage{
insert_tags(string, insertions)
}
\description{
inserts html tags into a string for a data.tree
}
\examples{
string = "Hello World"
insertions <- c("<b>" = 3, "</b>" = 6)
insert_tags(string, insertions)
}

