% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlFormat.R
\name{remove_tags}
\alias{remove_tags}
\title{removes html tags from strings}
\usage{
remove_tags(string, pattern = "<[^(<|>)]*>")
}
\arguments{
\item{string}{input string}

\item{pattern}{pattern for removal}
}
\description{
removes html tags from strings
}
\examples{
remove_tags("hello <b>world</b>")
}

