% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlFormat.R
\name{trim_tag}
\alias{trim_tag}
\title{Trims a tag surrounded by text to only tag}
\usage{
trim_tag(string)
}
\arguments{
\item{string}{input string containing tag and possibly other characters eg: 3e</b>gg}
}
\description{
Trims a tag surrounded by text to only tag
}
\examples{
trim_tag("hello <b> world")
}

